% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binSearchC}
\alias{binSearchC}
\title{Find an integer in an ordered vector using binary search.}
\usage{
binSearchC(i, v)
}
\arguments{
\item{i}{Integer to be found in vector.}

\item{v}{IntegerVector ordered increasingly in which \code{i} is to be found.}
}
\value{
boolean indicating whether \code{v} contains \code{i}.
}
\description{
Find an integer in an ordered vector using binary search.
}
\examples{
i = 42; v = 0:41
binSearchC(i,v)
}
\seealso{
\code{\link{match}}
}

