% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{filterAndNormalizeFront}
\alias{filterAndNormalizeFront}
\title{Filter and normalization of a paretro front}
\usage{
filterAndNormalizeFront(front, minimumdev, maximumdev, minimumconn, maximumconn)
}
\arguments{
\item{Minimumdev}{minimum value for deviation}

\item{Maximumdev}{maximum value for deviation}

\item{Minimumconn}{Minimum value for connectivity}

\item{Maximumconn}{Maximum value for connectivity}
}
\description{
Filter and normalization of a paretro front
}

