% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanVectorOfMatrixC}
\alias{meanVectorOfMatrixC}
\title{Calculate the mean vector for a matrix by computing the mean
for each column. This is an alternative to apply(m,2,mean).}
\usage{
meanVectorOfMatrixC(m)
}
\arguments{
\item{m}{NumericMatrix for which to calculate the mean vector.}
}
\value{
mean NumericVector containing the coordinates of the mean point of \code{m}.
}
\description{
Calculate the mean vector for a matrix by computing the mean
for each column. This is an alternative to apply(m,2,mean).
}
\examples{
m = matrix(1:25,ncol=5)
meanVectorOfMatrixC(m)
}
\seealso{
\code{\link{mean}}
}

