% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{midpoint}
\alias{midpoint}
\title{Calculate the (floor) mid of two integers.}
\usage{
midpoint(a, b)
}
\arguments{
\item{a}{First integer.}

\item{b}{Second integer.}
}
\value{
Floor mid of \code{a} and \code{b}.
}
\description{
Calculate the (floor) mid of two integers.
}

