% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mutateC}
\alias{mutateC}
\title{Mutation of individuals in nearest neighbour space}
\usage{
mutateC(sol, nn, neighborhood, randomVectorMutation,
  randomVectorNeighborSelection)
}
\arguments{
\item{sol:}{Solution to mutate}

\item{nn:}{Nearest neighbours matrix}

\item{neighborhood:}{Vector containing likelihood that an edge is mutated}

\item{randomVectorMutation:}{Comparison vector for neighborhood}

\item{randomVectorNeighborSelection:}{One of L nearest neighbors that is selected in case edge is mutated}
}
\description{
Mutation of individuals in nearest neighbour space
}

