% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutateMock.R
\name{mutateMock}
\alias{mutateMock}
\title{mutateMock Mutation operator for the MOCK algorithm}
\usage{
mutateMock(param, nn, neighborhoodMutationMatrix, L)
}
\arguments{
\item{param}{Representation of one cluster solution}

\item{nn}{Nearest neighbour matrix}

\item{neighborhoodMutationMatrix}{Matrix which contains the "costs" for each link i->j}
}
\value{
mutatedParam A mutated representatipn of one cluster solution
}
\description{
mutateMock Mutation operator for the MOCK algorithm
}

