% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{printAttainmentScore}
\alias{printAttainmentScore}
\title{printClusterSolution}
\usage{
printAttainmentScore(sol)
}
\arguments{
\item{sol}{Solution returned by MOCK function.}
}
\value{
plot Attainment plot
}
\description{
Plots the attainment score for a distinct solution returned by the mock function.
}
\details{
#'
}
\author{
Dennis Assenmacher
}

