% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{printClusterSolution}
\alias{printClusterSolution}
\title{printClusterSolution}
\usage{
printClusterSolution(solution, index, data, columns)
}
\arguments{
\item{solution}{Solution returned by MOCK function.}

\item{index}{Index or vector of indices of MOCK solution/s.}

\item{data}{Original input data as a matrix or dataframe.}

\item{columns}{Determines the column count of the plot.}
}
\description{
Draws a set of MOCK-Solutions in searchspace.
}
\details{
#'
}
\author{
Dennis Assenmacher
}

