% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{printMockRAndMocktoolFronts}
\alias{printMockRAndMocktoolFronts}
\title{PrintCurrentSolutions}
\usage{
printMockRAndMocktoolFronts(mockSolution, mockToolSolution, absolute = "T",
  dataset = "")
}
\arguments{
\item{sol}{Solution returned by MOCK function.}
}
\description{
Visualizes a 2-dimensional solution of PESA-II.
}

