% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{printParetoFront}
\alias{printParetoFront}
\title{PrintCurrentSolutions}
\usage{
printParetoFront(sol, method = "benchmark", maxClusters = 25,
  labelType = "index", gridIntervals = 0.1, dataset = "",
  markBestsolutions = T)
}
\arguments{
\item{sol}{Solution returned by MOCK function.}

\item{method}{Determines how the graph should be plotted. Options are: "benchmark","fronts","normalized","unnormalized".}

\item{maxClusters}{Determines which solutions should be visualized based on the cluster count.}

\item{labelType}{Determines whether the index of the cluster solution should be displayed within the plot.}

\item{dataset}{Name of the dataset used in order to print the title.}

\item{markBestsolutions}{Determines whether the best soultion should be displayed or not.}
}
\value{
ggplot to be created
}
\description{
Visualizes a 2-dimensional solution of PESA-II.
}

