% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resizeList}
\alias{resizeList}
\title{Resize an integervector to a given size.}
\usage{
resizeList(v, size)
}
\arguments{
\item{v}{IntegerVector to be resized. All elements with an index greather than
\code{size} are omitted.}

\item{size}{Integer determining the new seize of \code{v}.}
}
\value{
returnVector Resized IntegerVector \code{v}.
}
\description{
Resize an integervector to a given size.
}

