% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{selectRowsOfDatasetCInteger}
\alias{selectRowsOfDatasetCInteger}
\title{Select rows of a matrix based on a vector.}
\usage{
selectRowsOfDatasetCInteger(data, v)
}
\arguments{
\item{data}{A NumericMatrix on which a vector is to be applied.}

\item{v}{A vector that is applied on \code{data}.}
}
\value{
dataSelection NumericMatrx containing only the selected
rows of \code{data}.
}
\description{
Select rows of a matrix based on a vector.
}
\examples{
data = matrix(1:25,ncol=5); v = c(1,4)
selectRowsOfDatasetC(data,v)
}
\seealso{
base::Extract
}

