% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichC}
\alias{whichC}
\title{Zero-based Rcpp implementation of \code{which} for integer vectors.}
\usage{
whichC(x, v)
}
\arguments{
\item{x}{Integer to be found.}

\item{v}{IntegerVector in which to find \code{x}.}
}
\value{
indexVector IntegerVector containing all zero-based indexes of
appearences of \code{x} in \code{v}.
}
\description{
Zero-based Rcpp implementation of \code{which} for integer vectors.
}
\examples{
x = 42; v = 40:44
whichC(x,v)
}
\seealso{
\code{\link{which}}
}

